Move EvaluationContext to new file and add helper
